+++ django_flexquery/flexquery.py
@@ -5,30 +5,11 @@ model managers in a reusable fashion using Q objects.
 
 import inspect
 
-from django.core.exceptions import ImproperlyConfigured
 from django.db.models import Manager as _Manager, QuerySet as _QuerySet
 from django.utils.decorators import classonlymethod
 
 
-class FlexQueryType(type):
-    """
-    Custom metaclass for FlexQuery that implements the descriptor pattern.
-    """
-
-    def __get__(cls, instance, owner):
-        if cls.func is not None and isinstance(instance, (_Manager, _QuerySet)):
-            # Return a new FlexQuery instance bound to the holding object
-            return cls(instance)
-        # Otherwise, preserve the FlexQuery type
-        return cls
-
-    def __repr__(cls):
-        if cls.func is not None:
-            return "<type %s %r>" % (cls.__name__, cls.func.__name__)
-        return super().__repr__()
-
-
-class FlexQuery(metaclass=FlexQueryType):
+class FlexQuery:
     """
     Flexibly provides model-specific query constraints as an attribute of Manager
     or QuerySet objects.
@@ -45,25 +26,19 @@ class FlexQuery(metaclass=FlexQueryType):
     func = None
 
     def __call__(self, *args, **kwargs):
-        """Filters the base QuerySet using the provided function, relaying arguments.
+        """Filters the base queryset using the provided function, relaying arguments.
 
         :returns QuerySet:
         """
         return self.base.filter(self.call_bound(*args, **kwargs))
 
     def __init__(self, base):
-        """Binds a FlexQuery instance to the given base to perform its filtering on.
+        """Binds a FlexQuery sub-type to the given base to perform its filtering on.
 
         :param base: instance to use as base for filtering
         :type  base: Manager | QuerySet
-        :raises ImproperlyConfigured:
-            if this FlexQuery type wasn't created by the from_q() classmethod
-        :raises TypeError: if base is of unsupported type
+        :raises TypeError: if ``base`` is of unsupported type
         """
-        if self.func is None:
-            raise ImproperlyConfigured(
-                "Use the from_q() classmethod to create a FlexQuery sub-type."
-            )
         if not isinstance(base, (_Manager, _QuerySet)):
             raise TypeError(
                 "Can only bind %s to a Manager or QuerySet, but not to %r."
@@ -86,9 +61,9 @@ class FlexQuery(metaclass=FlexQueryType):
         return self.call_bound(*args, **kwargs)
 
     def call_bound(self, *args, **kwargs):
-        """Calls the provided Q function with self.base.all() as first argument.
+        """Calls the provided function with ``self.base.all()`` as first argument.
 
-        This may be overloaded if arguments need to be preprocessed in some way
+        This may be overwritten if arguments need to be preprocessed in some way
         before being passed to the custom function.
 
         :returns Q:
@@ -97,21 +72,46 @@ class FlexQuery(metaclass=FlexQueryType):
         return self.func(self.base.all(), *args, **kwargs)
 
     @classonlymethod
-    def from_q(cls, func):
-        """Creates a FlexQuery type from a Q function.
+    def from_func(cls, func):
+        """Creates a ``FlexQuery`` sub-type from a function.
 
-        :param func: callable taking a base QuerySet and returning a Q object
-        :type  func: callable
+        :param func: function taking a base ``QuerySet`` and returning a ``Q`` object
+        :type  func: function
         :returns FlexQueryType:
+        :raises TypeError: if ``func`` is no function
         """
-        return type(cls)(cls.__name__, (cls,), {"func": staticmethod(func)})
+        if not inspect.isfunction(func):
+            raise TypeError(
+                "Can only create FlexQuery from function, but %s was given."
+                % func.__class__.__name__
+            )
+        return InitializedFlexQueryType(
+            cls.__name__, (cls,), {"func": staticmethod(func)}
+        )
+
+
+class InitializedFlexQueryType(type):
+    """
+    Custom metaclass implementing the descriptor pattern for sub-types of ``FlexQuery``
+    with function attached.
+    """
+
+    def __get__(cls, instance, owner):
+        if isinstance(instance, (_Manager, _QuerySet)):
+            # Return a new FlexQuery instance bound to the holding object
+            return cls(instance)
+        # Otherwise, preserve the FlexQuery type
+        return cls
+
+    def __repr__(cls):
+        return "<type %s %r>" % (cls.__name__, cls.func.__name__)
 
 
 class Manager(_Manager):
     """
-    Use this Manager class' from_queryset() method if you want to derive a Manager from
-    a QuerySet that has FlexQuery's defined. If Django's native Manager.from_queryset()
-    was used instead, all FlexQuery's would be lost.
+    Use this class' ``from_queryset`` method if you want to derive a ``Manager``
+    from a ``QuerySet`` with ``FlexQuery`` members. If Django's native
+    ``Manager.from_queryset`` was used instead, those members would be lost.
     """
 
     @classmethod
@@ -120,7 +120,7 @@ class Manager(_Manager):
         methods.update(
             inspect.getmembers(
                 queryset_class,
-                predicate=lambda member: isinstance(member, FlexQueryType)
+                predicate=lambda member: isinstance(member, InitializedFlexQueryType)
                 and not getattr(member, "queryset_only", None),
             )
         )
@@ -129,8 +129,8 @@ class Manager(_Manager):
 
 class QuerySet(_QuerySet):
     """
-    Adds support for deriving a Manager from QuerySet via as_manager(), preserving
-    the FlexQuery's.
+    Adds support for deriving a ``Manager`` from a ``QuerySet`` class via
+    ``as_manager``, preserving ``FlexQuery`` members.
     """
 
     @classmethod