+++ CHANGELOG.md
@@ -6,11 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## 2.1.0 - Unreleased
+## 3.0.0 - Unreleased
 ### Added
 * Added `django_flexquery.contrib.for_user.ForUserFlexQuery`
 ### Changed
 * Changed directory structure of unittests
+* Renamed ``FlexQuery.from_q`` to ``from_func``
+* ``FlexQuery`` super-types (those without function attached) don't have a custom
+  metaclass anymore. This has no effect on the API.
 
 
 ## 2.0.0 - 2019-08-06